+++ src/java/src/main/java/org/wololo/flatgeobuf/GeometryOffsets.java
@@ -6,6 +6,6 @@ public class GeometryOffsets {
     public int[] lengths = null;
     public int endsOffset = 0;
     public int lengthsOffset = 0;
-    public byte type = 0;
+    public int type = 0;
     public GeometryOffsets[] gos = null;
 }
